adjusted signals for paper conformity, removed + adjusted power meters
